+++ resolwe_bio/processes/reads_processing/cutadapt_3prime.py
@@ -23,7 +23,7 @@ class Cutadapt3Prime(Process):
     slug = 'cutadapt-3prime-single'
     name = "Cutadapt (3' mRNA-seq, single-end)"
     process_type = 'data:reads:fastq:single:cutadapt:'
-    version = '1.1.0'
+    version = '1.1.1'
     category = 'Other'
     scheduling_class = SchedulingClass.BATCH
     entity = {'type': 'sample'}
@@ -96,7 +96,9 @@ class Cutadapt3Prime(Process):
     def run(self, inputs, outputs):
         """Run analysis."""
         # Get input reads file name (for the first of the possible multiple lanes)
-        name = os.path.basename(inputs.reads.fastq[0].path).strip('.fastq.gz')
+        reads_path = os.path.basename(inputs.reads.fastq[0].path)
+        assert reads_path.endswith('.fastq.gz')
+        name = reads_path[:-9]
         # Concatenate multi-lane read files
         (Cmd['cat'][[reads.path for reads in inputs.reads.fastq]] > 'input_reads.fastq.gz')()